+++ src/meegsim/utils.py
@@ -211,3 +211,10 @@ def vertices_to_mne(vertices, src):
         packed_vertices[src_idx] = src_vertno
 
     return packed_vertices
+
+
+def _hemi_to_index(hemi):
+    """
+    Get the index of the hemisphere (0 for lh, 1 for rh).
+    """
+    return ["lh", "rh"].index(hemi)