+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A desired level of white noise can be added in sensor space to model measurement
 noise ([#58](https://github.com/ctrltz/meegsim/pull/58))
+- A possibility to plot the source configuration ([#59](https://github.com/ctrltz/meegsim/pull/59))
 - Adjustment of global (all signal vs. all noise sources) SNR ([#64](https://github.com/ctrltz/meegsim/pull/64))
 - Adjustment of the standard deviation of source activity ([#66](https://github.com/ctrltz/meegsim/pull/66))